requirements.txt
dm-haiku==0.0.6
jax==0.3.4
jaxlib==0.3.2
h5py==3.6.0
requests==2.27.1
tqdm==4.64.0
yapf==0.32.0
